Crate traffic_signal_data

A representation of traffic signal configuration that references OpenStreetMap IDs and is hopefully robust over minor edits over time.

Structs

DirectedRoad
A road segment connecting two intersections, and a direction along the segment.

Phase
A traffic signal is in one phase at any time. The phase describes what movements are possible.

TrafficSignal

Turn
A movement through an intersection.

Enums

PhaseType
How long a phase lasts before moving to the next one.

Statics

DATA

Functions

load_all_data
Returns all traffic signal data compiled into this build, keyed by OSM node ID.